requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Resume Parsing
pdfplumber>=0.10.3
python-docx>=1.1.0

# NLP
spacy>=3.8.0

# OpenAI (optional, for AI matching)
openai>=1.30.1

# Email
# smtplib is built-in

# Scheduling (optional, for local use)
APScheduler>=3.10.4

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1

# Web API (backend)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
